Merge of master into new_platforms

+++ enclave/core/hostcalls.c
@@ -156,14 +156,14 @@ int oe_host_vfprintf(int device, const char* fmt, oe_va_list ap_)
     }
 
     /* If string was truncated, retry with correctly sized buffer */
-    if (n >= sizeof(buf))
+    if (n >= (int)sizeof(buf))
     {
-        if (!(p = oe_stack_alloc(n + 1)))
+        if (!(p = oe_stack_alloc((uint32_t)n + 1)))
             return -1;
 
         oe_va_list ap;
         oe_va_copy(ap, ap_);
-        n = oe_vsnprintf(p, n + 1, fmt, ap);
+        n = oe_vsnprintf(p, (size_t)n + 1, fmt, ap);
         oe_va_end(ap);
     }
